Extract bookmarks fetching into custom hook
