+++ packages/ui/src/components/domain/sidebar/NavSidebar.tsx
@@ -1,105 +1,42 @@
-import { useEffect, useState } from 'react'
 import { FiCompass, FiBookmark, FiSettings } from 'react-icons/fi'
 import { LuLogOut } from 'react-icons/lu'
 import { withRouter } from 'next/router'
-import { useUser } from '@supabase/auth-helpers-react'
 import Sidebar from '@/components/domain/sidebar'
 import Skeleton from '@/components/shared/Skeleton'
-import { Bookmark, PageInfo, useFilteredBookmarksQuery } from '@/graphql/generated/gql'
-import useSidebarCategories from '@/stores/useSidebarCategories'
+import useBookmarks from '@/hooks/useBookmarks'
 import Group from './Group'
 import Item from './Item'
 import MobileMenu from './MobileMenu'
 import Section from './Section'
-import useSidebarSync from './useSidebarSync'
-
-const renderFooter = () => (
-  <>
-    <Item Icon={FiSettings} text='Settings' path='/settings' />
-    <Item Icon={LuLogOut} text='Log out' path='/logout' />
-  </>
-)
-
-const PAGE_SIZE = 30
 
 const NavSidebar = () => {
-  const [bookmarks, setBookmarks] = useState<Bookmark[]>([])
-  const [pageInfo, setPageInfo] = useState<PageInfo>({
-    hasNextPage: true,
-    hasPreviousPage: false
-  })
-
-  const user = useUser()
-  const { counter } = useSidebarSync()
-  const { categoriesLength, setCategoriesLength } = useSidebarCategories()
-
-  // Fetch data from Supabase using generated Urql hook
-  const [{ data: urqlData, fetching }, refetch] = useFilteredBookmarksQuery({
-    variables: { userId: user?.id as string, first: PAGE_SIZE, after: pageInfo?.endCursor }
-  })
-
-  // Only update state when urql data changes
-  useEffect(() => {
-    if (urqlData) {
-      const edges = urqlData?.bookmarkCollection?.edges
-      setBookmarks(
-        [...bookmarks, ...(edges?.map(edge => edge.node) as Bookmark[])].reduceRight(
-          (acc, bookmark) =>
-            acc.some(b => b.project?.id === bookmark.project?.id) ? acc : [...acc, bookmark],
-          [] as Bookmark[]
-        )
-      )
-
-      if (pageInfo.hasNextPage) {
-        setPageInfo(urqlData?.bookmarkCollection?.pageInfo as PageInfo)
-      }
-    }
-  }, [urqlData, pageInfo])
-
-  // Refetch when bookmarks are added or removed
-  useEffect(() => {
-    if (counter > 0) {
-      setPageInfo({ hasNextPage: true, hasPreviousPage: false, endCursor: null })
-      refetch({
-        requestPolicy: 'network-only'
-      })
-    }
-  }, [counter])
-
-  useEffect(() => {
-    const newCategoriesLength = Array.from(new Set(bookmarks?.map(b => b.category))).length
-
-    // Only update categoriesLength if it has changed
-    if (newCategoriesLength !== categoriesLength) setCategoriesLength(newCategoriesLength)
-  }, [bookmarks])
-
-  const uniqueCategories = Array.from(new Set(bookmarks.map(bookmark => bookmark.category)))
+  const { bookmarks, categories, fetching, categoriesLength } = useBookmarks()
 
   return (
     <>
-      <Sidebar title='TruffleAI' footer={renderFooter()}>
+      <Sidebar title='TruffleAI'>
         <Section title='Overview'>
           <Item Icon={FiCompass} text='Trending projects' path='/' />
           <Item Icon={FiBookmark} text='All bookmarks' path='/bookmarks' />
         </Section>
 
         <Section title='Categories'>
-          {fetching && !urqlData ? (
+          {fetching && !bookmarks ? (
             <div className='mt-2 flex flex-col gap-3'>
               {Array.from(Array(categoriesLength).keys()).map(c => (
                 <Skeleton key={c} className='ml-5 h-6 !w-40' />
               ))}
             </div>
           ) : // Display categories as folders
-          uniqueCategories.length > 0 ? (
-            uniqueCategories
+          categories.length > 0 ? (
+            categories
               .sort((a, b) => (a ?? '').localeCompare(b ?? ''))
               .map(category => (
                 <div key={category}>
                   <Group
                     key={category}
-                    text={category as string}
-                    path={`/compare/${encodeURIComponent(category as string)}`}
+                    text={category}
+                    path={`/compare/${encodeURIComponent(category)}`}
                     bookmarks={bookmarks.filter(bookmark => bookmark.category === category)}
                   />
                 </div>
@@ -108,6 +45,11 @@ const NavSidebar = () => {
             <p className='py-2.5 pl-5 text-xs text-white/90'>No bookmarks yet</p>
           )}
         </Section>
+
+        <div className='border-t border-solid border-white/5 p-2'>
+          <Item Icon={FiSettings} text='Settings' path='/settings' />
+          <Item Icon={LuLogOut} text='Log out' path='/logout' />
+        </div>
       </Sidebar>
 
       <MobileMenu title='TruffleAI' bookmarks={bookmarks} />
