+++ packages/ui/src/hooks/useBookmarksSync.ts
@@ -5,9 +5,9 @@ interface SidebarSync {
   sync: () => void
 }
 
-const useSidebarSync = create<SidebarSync>(set => ({
+const useBookmarksSync = create<SidebarSync>(set => ({
   counter: 0,
   sync: () => set(state => ({ counter: state.counter + 1 }))
 }))
 
-export default useSidebarSync
+export default useBookmarksSync
