+++ packages/ui/src/hooks/useSidebarCategories.ts
